tests/krate/search: Move tests into `routes` submodules
