+++ src/tests/all.rs
@@ -38,6 +38,7 @@ mod krate;
 mod metrics;
 mod not_found_error;
 mod owners;
+mod pagination;
 mod read_only_mode;
 mod record;
 mod routes;
