+++ src/tests/krate/mod.rs
@@ -1,6 +1,5 @@
 mod following;
 mod publish;
-mod search;
 mod summary;
 mod versions;
 mod yanking;
