+++ src/tests/routes/crates/mod.rs
@@ -1,5 +1,6 @@
 pub mod downloads;
 mod following;
+mod list;
 mod new;
 pub mod owners;
 mod read;
